WIP integrate SQL query generation
