+++ src/lib.rs
@@ -1,8 +1,9 @@
-mod struct_builder;
+pub use column::*;
 pub use struct_builder::*;
+pub use sql::*;
+pub use types::*;
 
 mod column;
-pub use column::*;
-
+mod struct_builder;
+mod sql;
 mod types;
-pub use types::*;
